+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/ArgMode.java
@@ -2,10 +2,8 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
-import java.math.BigInteger;
-
 public interface ArgMode {
 
-    BigInteger getValue(Context context, int pointer);
+    long getValue(Context context, int pointer);
 
 }
