+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/PositionArgMode.java
@@ -2,14 +2,12 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
-import java.math.BigInteger;
-
 public class PositionArgMode implements ArgMode {
 
     @Override
-    public BigInteger getValue(Context context, int pointer) {
-        BigInteger[] code = context.getCode();
-        int address = code[pointer].intValue();
+    public long getValue(Context context, int pointer) {
+        long[] code = context.getCode();
+        int address = (int) code[pointer];
         return code[address];
     }
 }
