+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/opcode/OpCode3.java
@@ -2,19 +2,17 @@ package com.martak.adventofcode.intcodecomputer.model.opcode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
-import java.math.BigInteger;
-
 public class OpCode3 implements OpCode {
 
     @Override
     public Context execute(Context context) {
-        BigInteger[] code = context.getCode();
+        long[] code = context.getCode();
         int pointer = context.getPointer();
-        int address = code[pointer + 1].intValue();
-        BigInteger[] inputs = context.getInputs();
+        int address = (int) code[pointer + 1];
+        long[] inputs = context.getInputs();
         code[address] = inputs[0];
         if (inputs.length > 1) {
-            inputs = new BigInteger[]{inputs[1]};
+            inputs = new long[]{inputs[1]};
         }
         return new Context(pointer + 2, code, inputs, context.getOutput(), context.getRelativeBase());
     }
