+++ src/test/java/com/martak/adventofcode/amplificationcircuit/PhaseSequenceGeneratorTest.java
@@ -4,8 +4,6 @@ package com.martak.adventofcode.amplificationcircuit;
 import org.assertj.core.api.Assertions;
 import org.testng.annotations.Test;
 
-import java.math.BigInteger;
-
 
 public class PhaseSequenceGeneratorTest {
 
@@ -15,7 +13,7 @@ public class PhaseSequenceGeneratorTest {
         int min = 0;
         int max = 4;
         // when
-        BigInteger[][] actual = PhaseSequenceGenerator.getAllCombinations(min, max);
+        long[][] actual = PhaseSequenceGenerator.getAllCombinations(min, max);
         // then
         Assertions.assertThat(actual.length).isEqualTo(120);
     }
